+++ src/main/java/club/iananderson/seasonhud/SeasonHUD.java
@@ -9,22 +9,25 @@ import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.server.ServerStartingEvent;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
+import net.minecraftforge.fml.InterModComms;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.config.ModConfig;
 import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
+import net.minecraftforge.fml.event.lifecycle.InterModEnqueueEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import top.theillusivec4.curios.api.SlotTypeMessage;
 
 // The value here should match an entry in the META-INF/mods.toml file
 @Mod(SeasonHUD.MODID)
-public class  SeasonHUD{
+public class SeasonHUD{
 
-    // Define mod id in a common place for everything to reference
+    // Define mod id in an impl place for everything to reference
     public static final String MODID = "seasonhud";
     // Directly reference a slf4j logger
     //private static final Logger LOGGER = LogUtils.getLogger();
@@ -43,7 +46,9 @@ public class  SeasonHUD{
         IEventBus modEventBus = FMLJavaModLoadingContext.get().getModEventBus();
 
         // Register the commonSetup method for modloading
+        modEventBus.addListener(this::enqueue);
         modEventBus.addListener(this::commonSetup);
+
 
         // Register the Deferred Register to the mod event bus so blocks get registered
         BLOCKS.register(modEventBus);
@@ -56,12 +61,19 @@ public class  SeasonHUD{
         // Register ourselves for server and other game events we are interested in
         MinecraftForge.EVENT_BUS.register(this);
     }
+    public void enqueue(InterModEnqueueEvent event) {
+            LOGGER.info("Talking to Curios");
+            InterModComms.sendTo("curios", SlotTypeMessage.REGISTER_TYPE, () -> new SlotTypeMessage.Builder("charm")
+                    .size(1)
+                    .build());
+        }
 
     private void commonSetup(final FMLCommonSetupEvent event) {
-        // Some common setup code
+        // Some impl setup code
         LOGGER.info("HELLO FROM COMMON SETUP");
         LOGGER.info("DIRT BLOCK >> {}", ForgeRegistries.BLOCKS.getKey(Blocks.DIRT));
     }
+
 
     // You can use SubscribeEvent and let the Event Bus discover methods to call
     @SubscribeEvent