+++ src/main/java/club/iananderson/seasonhud/client/minimaps/JourneyMap.java
@@ -18,14 +18,11 @@ import net.minecraftforge.fml.ModList;
 
 import java.util.ArrayList;
 
-import static club.iananderson.seasonhud.client.Calendar.calendar;
+import static club.iananderson.seasonhud.impl.sereneseasons.Calendar.calendar;
 import static club.iananderson.seasonhud.config.Config.enableMod;
-import static club.iananderson.seasonhud.data.CurrentSeason.*;
+import static club.iananderson.seasonhud.impl.sereneseasons.CurrentSeason.*;
 import static club.iananderson.seasonhud.SeasonHUD.MODID;
 
-/*Todo
-    * Need to switch names over to translatable ones
- */
 
 public class JourneyMap {