+++ src/main/java/club/iananderson/seasonhud/impl/sereneseasons/Calendar.java
@@ -1,4 +1,4 @@
-package club.iananderson.seasonhud.client;
+package club.iananderson.seasonhud.impl.sereneseasons;
 
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.player.LocalPlayer;
@@ -7,15 +7,15 @@ import net.minecraft.world.item.Item;
 import sereneseasons.api.SSItems;
 
 import static club.iananderson.seasonhud.config.Config.needCalendar;
-
+import static club.iananderson.seasonhud.impl.curios.CuriosCalendar.isInCharmSlot;
 
 public class Calendar {
     public static boolean invCalendar;
 
-    public static boolean calendar(){
-        Item calendar = SSItems.CALENDAR.get();
+    public static Item calendar = SSItems.CALENDAR.get();
 
-        if(needCalendar.get()) {
+    public static boolean calendar() {
+        if (needCalendar.get()) {
             Minecraft mc = Minecraft.getInstance();
             LocalPlayer player = mc.player;
 
@@ -23,22 +23,18 @@ public class Calendar {
                 Inventory inv = player.getInventory();
                 int slot = findCalendar(inv, calendar);
 
-                invCalendar = (slot >= 0);
-
-//                PoseStack seasonStack = new PoseStack();
-//                mc.font.drawShadow(seasonStack, String.valueOf(invCalendar), 50, 50, 0xffffffff);
+                invCalendar = (slot >= 0) || isInCharmSlot(); //TODO crashing when selecting Need Calendar in menu
 
             }
 
         }
         else invCalendar = true;
-//        System.out.println(invCalendar);
 
         return invCalendar;
     }
 
     private static int findCalendar(Inventory inv, Item item) {
-        for(int i = 0; i < inv.items.size(); ++i) {
+        for (int i = 0; i < inv.items.size(); ++i) {
             if ((!inv.items.get(i).isEmpty() && inv.items.get(i).is(item))
                     || (!inv.offhand.get(0).isEmpty() && inv.offhand.get(0).is(item))) {
                 return i;
@@ -47,6 +43,4 @@ public class Calendar {
         return -1;
     }
 
-
 }
-